// src/app/peliculas/interfaces/pelicula.ts
// Generated by https://quicktype.io

export interface CachePeliculaStore {
    pelicula:   Pelicula[];
}

export interface Pelicula {
    id:            number;
    nombre:        string;
    rutaImagen:    string;
    descripcion:   string;
    duracion:      number;
    clasificacion: number;
    fechaCreacion: string;
    categoriaId:   number;
}